refactor(core): extract speech tests to sibling file
Move the single inline `#[cfg(test)] mod tests` block out of
primer-core/src/speech.rs into a sibling `speech/tests.rs`.

speech.rs: 625 -> 287 lines (-338). The test module is a descendant
so `use super::*` resolves to `speech` verbatim — no visibility edits,
zero churn. All 176 primer-core lib tests preserved; clippy + fmt clean.

Pure test-only relocation, no behaviour change.
